refactor: separate project component files into their own folders
