+++ src/main.ts
@@ -7,8 +7,8 @@ import './pictures/dictionary.png';
 import './pictures/favicon.png';
 import './pictures/profile-picture.jpg';
 import './pictures/git-logo.png';
-import { registration, logOut } from './components/auth';
-import createHeader from './components/create-header';
+import { registration, logOut } from './components/auth/auth';
+import createHeader from './components/create-header/create-header';
 
 createHeader('Rs Lang', true);
